+++ ovs/driver.go
@@ -1,7 +1,9 @@
 package ovs
 
 import (
+	"encoding/json"
 	"fmt"
+	"net/http"
 	"os/exec"
 	"reflect"
 	"strings"
@@ -9,18 +11,66 @@ import (
 
 	"github.com/docker/docker/api/types"
 	networkplugin "github.com/docker/go-plugins-helpers/network"
-	"github.com/docker/libnetwork/iptables"
 	log "github.com/sirupsen/logrus"
 )
 
-type OFPortMap struct {
+type ContainerState struct {
+	Name       string
+	Id         string
 	OFPort     uint32
-	AddPorts   string
-	Mode       string
-	NetworkID  string
-	EndpointID string
-	Options    map[string]interface{}
-	Operation  string
+	MacAddress string
+	HostIP     string
+	Labels     map[string]string
+}
+
+type ExternalPortState struct {
+	Name   string
+	OFPort uint32
+}
+
+type NetworkState struct {
+	NetworkName       string
+	BridgeName        string
+	BridgeDpid        string
+	BridgeDpidUint    uint64
+	BridgeVLAN        uint
+	MTU               uint
+	Mode              string
+	AddPorts          string
+	Gateway           string
+	GatewayMask       string
+	FlatBindInterface string
+	UseDHCP           bool
+	Userspace         bool
+	NATAcl            string
+	Containers        map[string]ContainerState
+	ExternalPorts     map[string]ExternalPortState
+}
+
+type DovesnapOp struct {
+	NewNetworkState      NetworkState
+	NewStackMirrorConfig StackMirrorConfig
+	VethName             string
+	MacAddress           string
+	AddPorts             string
+	Mode                 string
+	NetworkID            string
+	EndpointID           string
+	Options              map[string]interface{}
+	Operation            string
+}
+
+type NotifyMsg struct {
+	NetworkState NetworkState
+	Type         string
+	Operation    string
+	Details      map[string]string
+}
+
+type NotifyMsgJson struct {
+	Version uint
+	Time    int64
+	Msg     NotifyMsg
 }
 
 type StackingPort struct {
@@ -47,26 +97,10 @@ type Driver struct {
 	mirrorBridgeIn          string
 	mirrorBridgeOut         string
 	networks                map[string]NetworkState
-	ofportmapChan           chan OFPortMap
+	dovesnapOpChan          chan DovesnapOp
+	notifyMsgChan           chan NotifyMsg
+	webResponseChan         chan string
 	stackMirrorConfigs      map[string]StackMirrorConfig
-}
-
-// NetworkState is filled in at network creation time
-// it contains state that we wish to keep for each network
-type NetworkState struct {
-	NetworkName       string
-	BridgeName        string
-	BridgeDpid        string
-	BridgeDpidInt     int
-	BridgeVLAN        int
-	MTU               int
-	Mode              string
-	AddPorts          string
-	Gateway           string
-	GatewayMask       string
-	FlatBindInterface string
-	UseDHCP           bool
-	Userspace         bool
 }
 
 func (d *Driver) createLoopbackBridge() error {
@@ -126,12 +160,7 @@ func (d *Driver) createStackingBridge() error {
 	for _, stackingInterface := range d.stackingInterfaces {
 		remoteDP, remotePort, localInterface := d.mustGetStackingInterface(stackingInterface)
 
-		ofport, _, err := d.addInternalPort(dpName, localInterface, 0)
-		if err != nil {
-			log.Debugf("Error attaching veth [ %s ] to bridge [ %s ]", localInterface, dpName)
-			return err
-		}
-		log.Infof("Attached veth [ %s ] to bridge [ %s ] ofport %d", localInterface, dpName, ofport)
+		ofport, _ := d.mustAddInternalPort(dpName, localInterface, 0)
 		stackingConfig += fmt.Sprintf("%s: {stack: ", remoteDP)
 		if d.stackPriority1 == remoteDP {
 			stackingConfig += "{priority: 1}, "
@@ -187,6 +216,7 @@ func (d *Driver) ReOrCreateNetwork(r *networkplugin.CreateNetworkRequest, operat
 	gateway, mask := mustGetGatewayIP(r)
 	useDHCP := mustGetUseDHCP(r)
 	useUserspace := mustGetUserspace(r)
+	natAcl := mustGetNATAcl(r)
 
 	if useDHCP {
 		if mode != "flat" {
@@ -203,7 +233,7 @@ func (d *Driver) ReOrCreateNetwork(r *networkplugin.CreateNetworkRequest, operat
 	ns := NetworkState{
 		BridgeName:        bridgeName,
 		BridgeDpid:        dpid,
-		BridgeDpidInt:     mustGetIntFromHexStr(dpid),
+		BridgeDpidUint:    mustGetUintFromHexStr(dpid),
 		BridgeVLAN:        vlan,
 		MTU:               mtu,
 		Mode:              mode,
@@ -213,77 +243,45 @@ func (d *Driver) ReOrCreateNetwork(r *networkplugin.CreateNetworkRequest, operat
 		FlatBindInterface: bindInterface,
 		UseDHCP:           useDHCP,
 		Userspace:         useUserspace,
+		NATAcl:            natAcl,
+		Containers:        make(map[string]ContainerState),
+		ExternalPorts:     make(map[string]ExternalPortState),
 	}
 
-	d.networks[r.NetworkID] = ns
-	d.stackMirrorConfigs[r.NetworkID] = d.getStackMirrorConfig(r)
-
 	if operation == "create" {
-		if err := d.initBridge(r.NetworkID, controller, dpid, add_ports, useUserspace); err != nil {
+		if err := d.initBridge(ns, controller, dpid, add_ports, useUserspace); err != nil {
 			panic(err)
 		}
 	}
 
-	createmap := OFPortMap{
-		OFPort:     0,
-		AddPorts:   add_ports,
-		Mode:       mode,
-		NetworkID:  r.NetworkID,
-		EndpointID: bridgeName,
-		Operation:  operation,
+	createMsg := DovesnapOp{
+		NewNetworkState:      ns,
+		NewStackMirrorConfig: d.getStackMirrorConfig(r),
+		AddPorts:             add_ports,
+		Mode:                 mode,
+		NetworkID:            r.NetworkID,
+		EndpointID:           bridgeName,
+		Operation:            operation,
 	}
 
-	d.ofportmapChan <- createmap
+	d.dovesnapOpChan <- createMsg
 	return err
 }
 
 func (d *Driver) DeleteNetwork(r *networkplugin.DeleteNetworkRequest) error {
 	log.Debugf("Delete network request: %+v", r)
-	// remove the bridge from the faucet config if it exists
-	ns := d.networks[r.NetworkID]
-	log.Debugf("Deleting Bridge %s", ns.BridgeName)
-
-	d.faucetconfrpcer.mustDeleteDp(ns.NetworkName)
-
-	if usingMirrorBridge(d) {
-		err := d.deletePatchPort(ns.BridgeName, mirrorBridgeName)
-		if err != nil {
-			log.Errorf("Unable to delete patch port to mirror bridge because: %v", err)
-		}
+	deleteMsg := DovesnapOp{
+		NetworkID: r.NetworkID,
+		Operation: "delete",
 	}
 
-	if usingStacking(d) {
-		_, stackDpName, _ := d.getStackDP()
-		err := d.deletePatchPort(ns.BridgeName, stackDpName)
-		if err != nil {
-			log.Errorf("Unable to delete patch port between bridges because: %v", err)
-		}
-		if usingStackMirroring(d) {
-			lbBridgeName := d.mustGetLoopbackDP()
-			err = d.deletePatchPort(ns.BridgeName, lbBridgeName)
-			if err != nil {
-				log.Errorf("Unable to delete patch port to loopback bridge: %v", err)
-			}
-		}
-	}
-
-	_, err := d.deleteBridge(ns.BridgeName)
-	if err != nil {
-		log.Errorf("Deleting bridge %s failed: %s", ns.BridgeName, err)
-		return err
-	}
-
-	delete(d.networks, r.NetworkID)
-	delete(d.stackMirrorConfigs, r.NetworkID)
+	d.dovesnapOpChan <- deleteMsg
 	return nil
 }
 
 func (d *Driver) CreateEndpoint(r *networkplugin.CreateEndpointRequest) (*networkplugin.CreateEndpointResponse, error) {
 	log.Debugf("Create endpoint request: %+v", r)
-	localVethPair := vethPair(truncateID(r.EndpointID))
-	log.Debugf("Create vethPair")
-	res := &networkplugin.CreateEndpointResponse{Interface: &networkplugin.EndpointInterface{MacAddress: localVethPair.Attrs().HardwareAddr.String()}}
-	log.Debugf("Attached veth %+v,", r.Interface)
+	res := &networkplugin.CreateEndpointResponse{Interface: &networkplugin.EndpointInterface{}}
 	return res, nil
 }
 
@@ -344,14 +342,8 @@ func (d *Driver) Join(r *networkplugin.JoinRequest) (*networkplugin.JoinResponse
 	log.Debugf("Join request: %+v", r)
 	localVethPair := vethPair(truncateID(r.EndpointID))
 	addVethPair(localVethPair)
+	macAddress := getMacAddr(localVethPair.PeerName)
 	ns := d.networks[r.NetworkID]
-	ofport, _, err := d.addInternalPort(ns.BridgeName, localVethPair.Name, 0)
-	if err != nil {
-		log.Debugf("Error attaching veth [ %s ] to bridge [ %s ]", localVethPair.Name, ns.BridgeName)
-		return nil, err
-	}
-	log.Infof("Attached veth [ %s ] to bridge [ %s ] ofport %d", localVethPair.Name, ns.BridgeName, ofport)
-
 	res := &networkplugin.JoinResponse{
 		InterfaceName: networkplugin.InterfaceName{
 			// SrcName gets renamed to DstPrefix + ID on the container iface
@@ -361,134 +353,136 @@ func (d *Driver) Join(r *networkplugin.JoinRequest) (*networkplugin.JoinResponse
 		Gateway: ns.Gateway,
 	}
 	log.Debugf("Join endpoint %s:%s to %s", r.NetworkID, r.EndpointID, r.SandboxKey)
-	addmap := OFPortMap{
-		OFPort:     ofport,
-		AddPorts:   "",
-		Mode:       "",
+	joinMsg := DovesnapOp{
+		VethName:   localVethPair.Name,
 		NetworkID:  r.NetworkID,
 		EndpointID: r.EndpointID,
 		Options:    r.Options,
-		Operation:  "add",
+		MacAddress: macAddress,
+		Operation:  "join",
 	}
-	d.ofportmapChan <- addmap
+	d.dovesnapOpChan <- joinMsg
 	return res, nil
 }
 
 func (d *Driver) Leave(r *networkplugin.LeaveRequest) error {
 	log.Debugf("Leave request: %+v", r)
-	portID := fmt.Sprintf(ovsPortPrefix + truncateID(r.EndpointID))
-	ns := d.networks[r.NetworkID]
-	ofport, err := d.ovsdber.getOfPortNumber(portID)
-	if err != nil {
-		log.Errorf("Unable to get ofport number from %s", portID)
-		return err
-	}
-	localVethPair := vethPair(truncateID(r.EndpointID))
-	delVethPair(localVethPair)
-	err = d.ovsdber.deletePort(ns.BridgeName, portID)
-	if err != nil {
-		log.Errorf("OVS port [ %s ] delete transaction failed on bridge [ %s ] due to: %s", portID, ns.BridgeName, err)
-		return err
-	}
-	log.Infof("Deleted OVS port [ %s ] from bridge [ %s ]", portID, ns.BridgeName)
-	log.Debugf("Leave %s:%s", r.NetworkID, r.EndpointID)
-	rmmap := OFPortMap{
-		OFPort:     ofport,
-		AddPorts:   "",
-		Mode:       "",
+	leaveMsg := DovesnapOp{
 		NetworkID:  r.NetworkID,
 		EndpointID: r.EndpointID,
-		Operation:  "rm",
+		Operation:  "leave",
 	}
-	d.ofportmapChan <- rmmap
+	d.dovesnapOpChan <- leaveMsg
 	return nil
 }
 
-func mergeInterfacesYaml(dpName string, intDpid int, description string, addInterfaces string) string {
-	return fmt.Sprintf("{dps: {%s: {dp_id: %d, description: OVS Bridge %s, interfaces: {%s}}}}",
-		dpName, intDpid, description, addInterfaces)
-}
-
-func mustHandleCreate(d *Driver, mapMsg OFPortMap) {
+func mustHandleDeleteNetwork(d *Driver, opMsg DovesnapOp) {
 	defer func() {
 		if rerr := recover(); rerr != nil {
-			log.Errorf("mustHandleCreate failed: %v", rerr)
+			log.Errorf("mustHandleDeleteNetwork failed: %v", rerr)
 		}
 	}()
 
-	log.Debugf("network ID: %s", mapMsg.NetworkID)
-	netInspect := d.dockerer.mustGetNetworkInspectFromID(mapMsg.NetworkID)
+	// remove the bridge from the faucet config if it exists
+	ns := d.networks[opMsg.NetworkID]
+	log.Infof("Deleting network ID %s bridge %s", opMsg.NetworkID, ns.BridgeName)
+
+	d.faucetconfrpcer.mustDeleteDp(ns.NetworkName)
+
+	if usingMirrorBridge(d) {
+		d.mustDeletePatchPort(ns.BridgeName, mirrorBridgeName)
+	}
+
+	if usingStacking(d) {
+		_, stackDpName, _ := d.getStackDP()
+		d.mustDeletePatchPort(ns.BridgeName, stackDpName)
+		if usingStackMirroring(d) {
+			lbBridgeName := d.mustGetLoopbackDP()
+			d.mustDeletePatchPort(ns.BridgeName, lbBridgeName)
+		}
+	}
+
+	d.mustDeleteBridge(ns.BridgeName)
+
+	delete(d.networks, opMsg.NetworkID)
+	delete(d.stackMirrorConfigs, opMsg.NetworkID)
+
+	d.notifyMsgChan <- NotifyMsg{
+		Type:         "NETWORK",
+		Operation:    "DELETE",
+		NetworkState: ns,
+	}
+}
+
+func mustHandleCreateNetwork(d *Driver, opMsg DovesnapOp) {
+	defer func() {
+		if rerr := recover(); rerr != nil {
+			log.Errorf("mustHandleCreateNetwork failed: %v", rerr)
+		}
+	}()
+
+	log.Debugf("network ID: %s", opMsg.NetworkID)
+	netInspect := d.dockerer.mustGetNetworkInspectFromID(opMsg.NetworkID)
 	inspectNs, err := getNetworkStateFromResource(&netInspect)
 	if err != nil {
 		panic(err)
 	}
-	ns := d.networks[mapMsg.NetworkID]
+
+	ns := opMsg.NewNetworkState
+	d.stackMirrorConfigs[opMsg.NetworkID] = opMsg.NewStackMirrorConfig
 	ns.NetworkName = inspectNs.NetworkName
-	d.networks[mapMsg.NetworkID] = ns
-	add_ports := mapMsg.AddPorts
+	d.networks[opMsg.NetworkID] = ns
+
+	add_ports := opMsg.AddPorts
 	add_interfaces := ""
 	if add_ports != "" {
 		for _, add_port_number_str := range strings.Split(add_ports, ",") {
 			add_port_number := strings.Split(add_port_number_str, "/")
 			add_port := add_port_number[0]
-			ofport, err := d.ovsdber.getOfPortNumber(add_port)
-			if err != nil {
-				panic(fmt.Errorf("Unable to get ofport number from %s", add_port))
-			}
-			add_interfaces += fmt.Sprintf("%d: {description: %s, native_vlan: %d},", ofport, "Physical interface "+add_port, ns.BridgeVLAN)
+			ofport := d.ovsdber.mustGetOfPortNumber(add_port)
+			add_interfaces += d.faucetconfrpcer.vlanInterfaceYaml(ofport, "Physical interface "+add_port, ns.BridgeVLAN, "")
+			ns.ExternalPorts[add_port] = ExternalPortState{Name: add_port, OFPort: ofport}
 		}
 	}
-	mode := mapMsg.Mode
+	mode := opMsg.Mode
 	if mode == "nat" {
-		add_interfaces += fmt.Sprintf("%d: {description: OVS Port for NAT, native_vlan: %d},", ofPortLocal, ns.BridgeVLAN)
+		add_interfaces += d.faucetconfrpcer.vlanInterfaceYaml(ofPortLocal, "OVS Port for NAT", ns.BridgeVLAN, ns.NATAcl)
+		ns.ExternalPorts[inspectNs.BridgeName] = ExternalPortState{Name: inspectNs.BridgeName, OFPort: ofPortLocal}
 	}
-	configYaml := mergeInterfacesYaml(ns.NetworkName, ns.BridgeDpidInt, ns.BridgeName, add_interfaces)
+	configYaml := d.faucetconfrpcer.mergeInterfacesYaml(ns.NetworkName, ns.BridgeDpidUint, ns.BridgeName, add_interfaces)
 	if usingMirrorBridge(d) {
 		log.Debugf("configuring mirror bridge port for %s", ns.BridgeName)
-		stackMirrorConfig := d.stackMirrorConfigs[mapMsg.NetworkID]
-		ofportNum, mirrorOfportNum, err := d.addPatchPort(ns.BridgeName, mirrorBridgeName, stackMirrorConfig.LbPort, 0)
-		if err != nil {
-			panic(err)
-		}
-		flowStr := fmt.Sprintf("priority=2,in_port=%d,actions=mod_vlan_vid:%d,output:1", mirrorOfportNum, ns.BridgeVLAN)
+		stackMirrorConfig := d.stackMirrorConfigs[opMsg.NetworkID]
+		ofportNum, mirrorOfportNum := d.mustAddPatchPort(ns.BridgeName, mirrorBridgeName, stackMirrorConfig.LbPort, 0)
+		flowStr := fmt.Sprintf("priority=2,in_port=%d,dl_vlan=0xffff,actions=mod_vlan_vid:%d,output:1", mirrorOfportNum, ns.BridgeVLAN)
 		mustOfCtl("add-flow", mirrorBridgeName, flowStr)
 		add_interfaces += fmt.Sprintf("%d: {description: mirror, output_only: true},", ofportNum)
-		configYaml = mergeInterfacesYaml(ns.NetworkName, ns.BridgeDpidInt, ns.BridgeName, add_interfaces)
+		ns.ExternalPorts[mirrorBridgeName] = ExternalPortState{Name: mirrorBridgeName, OFPort: ofportNum}
+		configYaml = d.faucetconfrpcer.mergeInterfacesYaml(ns.NetworkName, ns.BridgeDpidUint, ns.BridgeName, add_interfaces)
 	}
 	if usingStacking(d) {
 		_, stackDpName, err := d.getStackDP()
 		if err != nil {
 			panic(err)
 		}
-		ofportNum, ofportNumPeer, err := d.addPatchPort(ns.BridgeName, stackDpName, 0, 0)
-		if err != nil {
-			panic(err)
-		}
-		localDpYaml := fmt.Sprintf("%s: {dp_id: %d, description: %s, interfaces: {%s %d: {description: %s, stack: {dp: %s, port: %d}}}}",
+		ofportNum, ofportNumPeer := d.mustAddPatchPort(ns.BridgeName, stackDpName, 0, 0)
+		ns.ExternalPorts[stackDpName] = ExternalPortState{Name: stackDpName, OFPort: ofportNum}
+		localDpYaml := fmt.Sprintf("%s: {dp_id: %d, description: %s, interfaces: {%s %s}}",
 			ns.NetworkName,
-			ns.BridgeDpidInt,
+			ns.BridgeDpidUint,
 			"OVS Bridge "+ns.BridgeName,
 			add_interfaces,
-			ofportNum,
-			"Stack link to "+stackDpName,
+			d.faucetconfrpcer.stackInterfaceYaml(ofportNum, stackDpName, ofportNumPeer))
+		remoteDpYaml := fmt.Sprintf("%s: {interfaces: {%s}}",
 			stackDpName,
-			ofportNumPeer)
-		remoteDpYaml := fmt.Sprintf("%s: {interfaces: {%d: {description: %s, stack: {dp: %s, port: %d}}}}",
-			stackDpName,
-			ofportNumPeer,
-			"Stack link to "+ns.NetworkName,
-			ns.NetworkName,
-			ofportNum)
+			d.faucetconfrpcer.stackInterfaceYaml(ofportNumPeer, ns.NetworkName, ofportNum))
 		configYaml = fmt.Sprintf("{dps: {%s, %s}}", localDpYaml, remoteDpYaml)
 	}
 	d.faucetconfrpcer.mustSetFaucetConfigFile(configYaml)
 	if usingStackMirroring(d) {
 		lbBridgeName := d.mustGetLoopbackDP()
-		stackMirrorConfig := d.stackMirrorConfigs[mapMsg.NetworkID]
-		_, _, err = d.addPatchPort(ns.BridgeName, lbBridgeName, stackMirrorConfig.LbPort, 0)
-		if err != nil {
-			panic(err)
-		}
+		stackMirrorConfig := d.stackMirrorConfigs[opMsg.NetworkID]
+		d.mustAddPatchPort(ns.BridgeName, lbBridgeName, stackMirrorConfig.LbPort, 0)
 		d.faucetconfrpcer.mustSetRemoteMirrorPort(
 			ns.NetworkName,
 			stackMirrorConfig.LbPort,
@@ -496,6 +490,11 @@ func mustHandleCreate(d *Driver, mapMsg OFPortMap) {
 			stackMirrorConfig.RemoteDpName,
 			stackMirrorConfig.RemoteMirrorPort,
 		)
+	}
+	d.notifyMsgChan <- NotifyMsg{
+		Type:         "NETWORK",
+		Operation:    "CREATE",
+		NetworkState: ns,
 	}
 }
 
@@ -511,40 +510,34 @@ func mustGetPortMap(portMapRaw interface{}) (string, string, string) {
 	return hostPort, port, ipProto
 }
 
-func mustHandleAdd(d *Driver, mapMsg OFPortMap, OFPorts *map[string]OFPortContainer) {
+func mustHandleJoinContainer(d *Driver, opMsg DovesnapOp, OFPorts *map[string]OFPortContainer) {
 	defer func() {
 		if rerr := recover(); rerr != nil {
-			log.Errorf("mustHandleAdd failed: %v", rerr)
+			log.Errorf("mustHandleJoinContainer failed: %v", rerr)
 		}
 	}()
-	containerInspect, err := d.dockerer.getContainerFromEndpoint(mapMsg.EndpointID)
+	containerInspect, err := d.dockerer.getContainerFromEndpoint(opMsg.EndpointID)
 	if err != nil {
 		panic(err)
 	}
-	ns := d.networks[mapMsg.NetworkID]
+	ns := d.networks[opMsg.NetworkID]
 	pid := containerInspect.State.Pid
 	containerNetSettings := containerInspect.NetworkSettings.Networks[ns.NetworkName]
-
-	log.Infof("Adding %s (pid %d) on %s DPID %d OFPort %d to Faucet",
-		containerInspect.Name, pid, ns.BridgeName, ns.BridgeDpidInt, mapMsg.OFPort)
+	ofport, _ := d.mustAddInternalPort(ns.BridgeName, opMsg.VethName, 0)
+	log.Infof("Adding %s (pid %d) veth %s on %s DPID %d OFPort %d to Faucet",
+		containerInspect.Name, pid, opMsg.VethName, ns.BridgeName, ns.BridgeDpidUint, ofport)
 	log.Debugf("container network settings: %+v", containerNetSettings)
 
-	log.Debugf("%+v", mapMsg.Options[portMapOption])
+	log.Debugf("%+v", opMsg.Options[portMapOption])
 	hostIP := containerNetSettings.IPAddress
 	gatewayIP := containerNetSettings.Gateway
 
 	// Regular docker uses docker proxy, to listen on the configured port and proxy them into the container.
 	// dovesnap doesn't get to use docker proxy, so we listen on the configured port on the network's gateway instead.
-	for _, portMapRaw := range mapMsg.Options[portMapOption].([]interface{}) {
+	for _, portMapRaw := range opMsg.Options[portMapOption].([]interface{}) {
 		log.Debugf("adding portmap %+v", portMapRaw)
 		hostPort, port, ipProto := mustGetPortMap(portMapRaw)
-		iptables.Raw("-t", "nat", "-C", "DOCKER", "-p", ipProto, "-d", gatewayIP, "--dport", hostPort, "-j", "DNAT", "--to-destination", fmt.Sprintf("%s:%s", hostIP, port))
-		iptables.Raw("-t", "nat", "-A", "DOCKER", "-p", ipProto, "-d", gatewayIP, "--dport", hostPort, "-j", "DNAT", "--to-destination", fmt.Sprintf("%s:%s", hostIP, port))
-		iptables.Raw("-t", "nat", "-A", "OUTPUT", "-p", ipProto, "-d", gatewayIP, "--dport", hostPort, "-j", "DNAT", "--to-destination", fmt.Sprintf("%s:%s", hostIP, port))
-		iptables.Raw("-t", "nat", "-C", "POSTROUTING", "-p", ipProto, "-s", hostIP, "-d", hostIP, "--dport", port, "-j", "MASQUERADE")
-		iptables.Raw("-t", "nat", "-A", "POSTROUTING", "-p", ipProto, "-s", hostIP, "-d", hostIP, "--dport", port, "-j", "MASQUERADE")
-		iptables.Raw("-t", "filter", "-C", "DOCKER", "!", "-i", ns.BridgeName, "-o", ns.BridgeName, "-p", "tcp", "-d", hostIP, "--dport", port, "-j", "ACCEPT")
-		iptables.Raw("-t", "filter", "-A", "DOCKER", "!", "-i", ns.BridgeName, "-o", ns.BridgeName, "-p", "tcp", "-d", hostIP, "--dport", port, "-j", "ACCEPT")
+		mustAddGatewayPortMap(ns.BridgeName, ipProto, gatewayIP, hostIP, hostPort, port)
 	}
 
 	portacl := ""
@@ -552,17 +545,17 @@ func mustHandleAdd(d *Driver, mapMsg OFPortMap, OFPorts *map[string]OFPortContai
 	if ok && len(portacl) > 0 {
 		log.Infof("Set portacl %s on %s", portacl, containerInspect.Name)
 	}
-	add_interfaces := fmt.Sprintf("%d: {description: '%s', native_vlan: %d, acls_in: [%s]},",
-		mapMsg.OFPort, fmt.Sprintf("%s %s", containerInspect.Name, truncateID(containerInspect.ID)), ns.BridgeVLAN, portacl)
+	add_interfaces := d.faucetconfrpcer.vlanInterfaceYaml(
+		ofport, fmt.Sprintf("%s %s", containerInspect.Name, truncateID(containerInspect.ID)), ns.BridgeVLAN, portacl)
 
-	d.faucetconfrpcer.mustSetFaucetConfigFile(mergeInterfacesYaml(ns.NetworkName, ns.BridgeDpidInt, ns.BridgeName, add_interfaces))
+	d.faucetconfrpcer.mustSetFaucetConfigFile(d.faucetconfrpcer.mergeInterfacesYaml(ns.NetworkName, ns.BridgeDpidUint, ns.BridgeName, add_interfaces))
 
 	mirror, ok := containerInspect.Config.Labels["dovesnap.faucet.mirror"]
 	if ok && parseBool(mirror) {
 		log.Infof("Mirroring container %s", containerInspect.Name)
-		stackMirrorConfig := d.stackMirrorConfigs[mapMsg.NetworkID]
+		stackMirrorConfig := d.stackMirrorConfigs[opMsg.NetworkID]
 		if usingStackMirroring(d) || usingMirrorBridge(d) {
-			d.faucetconfrpcer.mustAddPortMirror(ns.NetworkName, uint32(mapMsg.OFPort), stackMirrorConfig.LbPort)
+			d.faucetconfrpcer.mustAddPortMirror(ns.NetworkName, ofport, stackMirrorConfig.LbPort)
 		}
 	}
 
@@ -570,6 +563,7 @@ func mustHandleAdd(d *Driver, mapMsg OFPortMap, OFPorts *map[string]OFPortContai
 
 	defaultInterface := "eth0"
 	udhcpcCmd := exec.Command("ip", "netns", "exec", containerInspect.ID, "/sbin/udhcpc", "-f", "-R", "-i", defaultInterface)
+	// TODO: If DHCP in use, need background process to obtain IP address.
 	if ns.UseDHCP {
 		err = udhcpcCmd.Start()
 		if err != nil {
@@ -587,58 +581,85 @@ func mustHandleAdd(d *Driver, mapMsg OFPortMap, OFPorts *map[string]OFPortContai
 		}
 	}
 	containerMap := OFPortContainer{
-		OFPort:           mapMsg.OFPort,
+		OFPort:           ofport,
 		containerInspect: containerInspect,
 		udhcpcCmd:        udhcpcCmd,
-		Options:          mapMsg.Options,
+		Options:          opMsg.Options,
 	}
-	(*OFPorts)[mapMsg.EndpointID] = containerMap
+	(*OFPorts)[opMsg.EndpointID] = containerMap
+	ns.Containers[opMsg.EndpointID] = ContainerState{
+		Name:       containerInspect.Name,
+		Id:         containerInspect.ID,
+		OFPort:     ofport,
+		MacAddress: opMsg.MacAddress,
+		HostIP:     hostIP,
+		Labels:     containerInspect.Config.Labels,
+	}
+
+	d.notifyMsgChan <- NotifyMsg{
+		Type:         "CONTAINER",
+		Operation:    "JOIN",
+		NetworkState: ns,
+		Details: map[string]string{
+			"name": containerInspect.Name,
+			"id":   containerInspect.ID,
+			"port": fmt.Sprintf("%d", ofport),
+			"mac":  opMsg.MacAddress,
+			"ip":   hostIP,
+		},
+	}
 }
 
-func mustHandleRm(d *Driver, mapMsg OFPortMap, OFPorts *map[string]OFPortContainer) {
+func mustHandleLeaveContainer(d *Driver, opMsg DovesnapOp, OFPorts *map[string]OFPortContainer) {
 	defer func() {
 		if rerr := recover(); rerr != nil {
-			log.Errorf("mustHandleRm failed: %v", rerr)
+			log.Errorf("mustHandleLeaveContainer failed: %v", rerr)
 		}
 	}()
-
-	containerMap := (*OFPorts)[mapMsg.EndpointID]
+	containerMap := (*OFPorts)[opMsg.EndpointID]
 	udhcpcCmd := containerMap.udhcpcCmd
 	if udhcpcCmd != nil {
 		log.Infof("Shutting down udhcpc")
 		udhcpcCmd.Process.Kill()
 		udhcpcCmd.Wait()
 	}
-	ns, have_network := d.networks[mapMsg.NetworkID]
-	if have_network {
-		log.Debugf("Removing port %d on %s from Faucet config", mapMsg.OFPort, ns.NetworkName)
-		d.faucetconfrpcer.mustDeleteDpInterface(ns.NetworkName, uint32(mapMsg.OFPort))
+	portID := fmt.Sprintf(ovsPortPrefix + truncateID(opMsg.EndpointID))
+	ofport := d.ovsdber.mustGetOfPortNumber(portID)
+	localVethPair := vethPair(truncateID(opMsg.EndpointID))
+	delVethPair(localVethPair)
 
-		containerNetSettings := containerMap.containerInspect.NetworkSettings.Networks[ns.NetworkName]
-		hostIP := containerNetSettings.IPAddress
-		gatewayIP := containerNetSettings.Gateway
-		for _, portMapRaw := range containerMap.Options[portMapOption].([]interface{}) {
-			log.Debugf("deleting portmap %+v", portMapRaw)
-			hostPort, port, ipProto := mustGetPortMap(portMapRaw)
-			iptables.Raw("-t", "nat", "-D", "DOCKER", "-p", ipProto, "-d", gatewayIP, "--dport", hostPort, "-j", "DNAT", "--to-destination", fmt.Sprintf("%s:%s", hostIP, port))
-			iptables.Raw("-t", "nat", "-D", "OUTPUT", "-p", ipProto, "-d", gatewayIP, "--dport", hostPort, "-j", "DNAT", "--to-destination", fmt.Sprintf("%s:%s", hostIP, port))
-			iptables.Raw("-t", "nat", "-D", "POSTROUTING", "-p", ipProto, "-s", hostIP, "-d", hostIP, "--dport", port, "-j", "MASQUERADE")
-			iptables.Raw("-t", "filter", "-D", "DOCKER", "!", "-i", ns.BridgeName, "-o", ns.BridgeName, "-p", "tcp", "-d", hostIP, "--dport", port, "-j", "ACCEPT")
-		}
+	ns := d.networks[opMsg.NetworkID]
+	d.ovsdber.mustDeletePort(ns.BridgeName, portID)
+	d.faucetconfrpcer.mustDeleteDpInterface(ns.NetworkName, ofport)
+
+	containerNetSettings := containerMap.containerInspect.NetworkSettings.Networks[ns.NetworkName]
+	hostIP := containerNetSettings.IPAddress
+	gatewayIP := containerNetSettings.Gateway
+	for _, portMapRaw := range containerMap.Options[portMapOption].([]interface{}) {
+		hostPort, port, ipProto := mustGetPortMap(portMapRaw)
+		mustDeleteGatewayPortMap(ns.BridgeName, ipProto, gatewayIP, hostIP, hostPort, port)
 	}
 
-	// The container will be gone by the time we query docker.
-	delete(*OFPorts, mapMsg.EndpointID)
+	delete(*OFPorts, opMsg.EndpointID)
+	delete(ns.Containers, opMsg.EndpointID)
+
+	d.notifyMsgChan <- NotifyMsg{
+		Type:         "CONTAINER",
+		Operation:    "LEAVE",
+		NetworkState: ns,
+		Details: map[string]string{
+			"name": containerMap.containerInspect.Name,
+			"id":   containerMap.containerInspect.ID,
+			"port": fmt.Sprintf("%d", ofport),
+		},
+	}
 }
 
 func reconcileOvs(d *Driver, allPortDesc *map[string]map[uint32]string) {
 	for id, ns := range d.networks {
 		stackMirrorConfig := d.stackMirrorConfigs[id]
 		newPortDesc := make(map[uint32]string)
-		err := scrapePortDesc(ns.BridgeName, &newPortDesc)
-		if err != nil {
-			continue
-		}
+		mustScrapePortDesc(ns.BridgeName, &newPortDesc)
 		addPorts := make(map[string]uint32)
 		d.ovsdber.parseAddPorts(ns.AddPorts, &addPorts)
 
@@ -660,6 +681,7 @@ func reconcileOvs(d *Driver, allPortDesc *map[string]map[uint32]string) {
 				}
 				log.Infof("removing non dovesnap port: %s %s %d %s", id, ns.BridgeName, ofport, desc)
 				d.faucetconfrpcer.mustDeleteDpInterface(ns.NetworkName, uint32(ofport))
+				delete(ns.ExternalPorts, desc)
 			}
 		} else {
 			log.Debugf("new portDesc for %s", ns.BridgeName)
@@ -682,11 +704,12 @@ func reconcileOvs(d *Driver, allPortDesc *map[string]map[uint32]string) {
 				continue
 			}
 			log.Infof("adding non dovesnap port: %s %s %d %s", id, ns.BridgeName, ofport, desc)
-			add_interfaces += fmt.Sprintf("%d: {description: %s, native_vlan: %d},", ofport, "Physical interface "+desc, ns.BridgeVLAN)
+			add_interfaces += d.faucetconfrpcer.vlanInterfaceYaml(ofport, "Physical interface "+desc, ns.BridgeVLAN, "")
+			ns.ExternalPorts[desc] = ExternalPortState{Name: desc, OFPort: ofport}
 		}
 
 		if add_interfaces != "" {
-			configYaml := mergeInterfacesYaml(ns.NetworkName, ns.BridgeDpidInt, ns.BridgeName, add_interfaces)
+			configYaml := d.faucetconfrpcer.mergeInterfacesYaml(ns.NetworkName, ns.BridgeDpidUint, ns.BridgeName, add_interfaces)
 			d.faucetconfrpcer.mustSetFaucetConfigFile(configYaml)
 		}
 
@@ -694,26 +717,37 @@ func reconcileOvs(d *Driver, allPortDesc *map[string]map[uint32]string) {
 	}
 }
 
-func consolidateDockerInfo(d *Driver) {
+func mustHandleNetworks(d *Driver) {
+	encodedMsg, err := json.Marshal(d.networks)
+	if err != nil {
+		panic(err)
+	}
+	d.webResponseChan <- fmt.Sprintf("%s", encodedMsg)
+}
+
+func (d *Driver) resourceManager() {
 	OFPorts := make(map[string]OFPortContainer)
 	AllPortDesc := make(map[string]map[uint32]string)
 
 	for {
 		select {
-		case mapMsg := <-d.ofportmapChan:
-			switch mapMsg.Operation {
+		case opMsg := <-d.dovesnapOpChan:
+			switch opMsg.Operation {
 			case "create":
-				mustHandleCreate(d, mapMsg)
-			case "add":
-				mustHandleAdd(d, mapMsg, &OFPorts)
-			case "rm":
-				mustHandleRm(d, mapMsg, &OFPorts)
+				mustHandleCreateNetwork(d, opMsg)
+			case "delete":
+				mustHandleDeleteNetwork(d, opMsg)
+			case "join":
+				mustHandleJoinContainer(d, opMsg, &OFPorts)
+			case "leave":
+				mustHandleLeaveContainer(d, opMsg, &OFPorts)
+			case "networks":
+				mustHandleNetworks(d)
 			default:
-				log.Errorf("Unknown consolidation message: %s", mapMsg)
+				log.Errorf("Unknown resource manager message: %s", opMsg)
 			}
-		default:
+		case <-time.After(time.Second * 3):
 			reconcileOvs(d, &AllPortDesc)
-			time.Sleep(3 * time.Second)
 		}
 	}
 }
@@ -730,7 +764,26 @@ func usingStackMirroring(d *Driver) bool {
 	return usingStacking(d) && len(d.stackMirrorInterface) > 1
 }
 
-func restoreNetworks(d *Driver) {
+func (d *Driver) notifier() {
+	for {
+		select {
+		case notifyMsg := <-d.notifyMsgChan:
+			log.Debugf("%+v", notifyMsg)
+			encodedMsg, err := json.Marshal(NotifyMsgJson{
+				Version: 1,
+				Time:    time.Now().Unix(),
+				Msg:     notifyMsg,
+			})
+			if err != nil {
+				panic(err)
+			}
+			// TODO: emit to UDS
+			log.Infof(fmt.Sprintf("%s", encodedMsg))
+		}
+	}
+}
+
+func (d *Driver) restoreNetworks() {
 	netlist := d.dockerer.mustGetNetworkList()
 	for id, _ := range netlist {
 		netInspect := d.dockerer.mustGetNetworkInspectFromID(id)
@@ -746,7 +799,25 @@ func restoreNetworks(d *Driver) {
 	}
 }
 
-func NewDriver(flagFaucetconfrpcServerName string, flagFaucetconfrpcServerPort int, flagFaucetconfrpcKeydir string, flagStackPriority1 string, flagStackingInterfaces string, flagStackMirrorInterface string, flagDefaultControllers string, flagMirrorBridgeIn string, flagMirrorBridgeOut string) *Driver {
+func (d *Driver) getWebResponse(w http.ResponseWriter, operation string) {
+	d.dovesnapOpChan <- DovesnapOp{Operation: operation}
+	response := <-d.webResponseChan
+	fmt.Fprintf(w, response)
+}
+
+func (d *Driver) handleNetworksWeb(w http.ResponseWriter, r *http.Request) {
+	d.getWebResponse(w, "networks")
+}
+
+func (d *Driver) runWeb(port int) {
+	http.HandleFunc("/networks", d.handleNetworksWeb)
+
+	if err := http.ListenAndServe(fmt.Sprintf(":%d", port), nil); err != nil {
+		panic(err)
+	}
+}
+
+func NewDriver(flagFaucetconfrpcServerName string, flagFaucetconfrpcServerPort int, flagFaucetconfrpcKeydir string, flagStackPriority1 string, flagStackingInterfaces string, flagStackMirrorInterface string, flagDefaultControllers string, flagMirrorBridgeIn string, flagMirrorBridgeOut string, flagStatusServerPort int) *Driver {
 	ensureDirExists(netNsPath)
 
 	stack_mirror_interface := strings.Split(flagStackMirrorInterface, ":")
@@ -767,7 +838,9 @@ func NewDriver(flagFaucetconfrpcServerName string, flagFaucetconfrpcServerPort i
 		mirrorBridgeIn:          flagMirrorBridgeIn,
 		mirrorBridgeOut:         flagMirrorBridgeOut,
 		networks:                make(map[string]NetworkState),
-		ofportmapChan:           make(chan OFPortMap, 2),
+		dovesnapOpChan:          make(chan DovesnapOp, 2),
+		notifyMsgChan:           make(chan NotifyMsg, 2),
+		webResponseChan:         make(chan string, 2),
 		stackMirrorConfigs:      make(map[string]StackMirrorConfig),
 	}
 
@@ -775,6 +848,8 @@ func NewDriver(flagFaucetconfrpcServerName string, flagFaucetconfrpcServerPort i
 	d.faucetconfrpcer.mustGetGRPCClient(flagFaucetconfrpcServerName, flagFaucetconfrpcServerPort, flagFaucetconfrpcKeydir)
 
 	d.ovsdber.waitForOvs()
+
+	go d.notifier()
 
 	if usingMirrorBridge(d) {
 		d.createMirrorBridge()
@@ -795,9 +870,11 @@ func NewDriver(flagFaucetconfrpcServerName string, flagFaucetconfrpcServerPort i
 		log.Warnf("No stacking interface defined, not stacking DPs or creating a stacking bridge")
 	}
 
-	restoreNetworks(d)
+	d.restoreNetworks()
 
-	go consolidateDockerInfo(d)
+	go d.resourceManager()
+
+	go d.runWeb(flagStatusServerPort)
 
 	return d
 }